+++ src/pages/profile/userProfile/profileInfo/ProfileInfo.presenter.jsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useContext, useEffect, useState } from "react";
+import React from "react";
 import {
   ProfileInfoFollowering,
   ProfileInfoFolloweringCount,
@@ -19,49 +19,17 @@ import {
 } from "./ProfileInfo.styles";
 import messageIcon from "../../../../img/icon-message-circle.svg";
 import shareIcon from "../../../../img/icon-share.svg";
-
 import profileImg from "../../../../img/ProfileImg.svg";
 import Button from "../../../../components/commons/button/Button";
-import { useNavigate, useParams } from "react-router-dom";
-import { customAxios } from "../../../../library/customAxios";
-import { AccountContext } from "../../../../context/AccountContext";
-
-export default function ProfileInfo({ userData }) {
-  const [isfollow, setIsfollow] = useState(userData.isfollow);
-  const [followCount, setFollowCount] = useState(userData.followerCount);
-  const navigate = useNavigate();
-  const { account } = useContext(AccountContext);
-  const params = useParams();
-  const userAccountname = params.userId;
-  function onClickButton(url) {
-    navigate(url);
-  }
-
-  // 좋은 방법이 아님 추후 수정예정
-  useEffect(()=>{
-    setFollowCount(userData.followerCount)
-  },[userData])
-
-  const onClickFollow = useCallback(async () => {
-    try {
-      await customAxios.post(`profile/${userAccountname}/follow`);
-      setIsfollow(true);
-      setFollowCount((prev) => prev + 1);
-    } catch (error) {
-      console.log(error);
-    }
-  }, [isfollow, followCount]);
-
-  const onClickUnfollow = useCallback(async () => {
-    try {
-      await customAxios.delete(`profile/${userAccountname}/unfollow`);
-      setIsfollow(false);
-      setFollowCount((prev) => prev - 1);
-    } catch (error) {
-      console.log(error);
-    }
-  }, [isfollow, followCount]);
-
+export default function ProfileInfoUI({
+  userData,
+  followCount,
+  account,
+  isfollow,
+  onClickButton,
+  onClickUnfollow,
+  onClickFollow,
+}) {
   return (
     <ProfileInfoWrapper>
       <ProfileInfoTitle className="a11y-hidden">프로필 정보</ProfileInfoTitle>
@@ -102,9 +70,7 @@ export default function ProfileInfo({ userData }) {
           <>
             <Button
               className="medium"
-              onClick={() =>
-                onClickButton(`/profile/myinfo/edit`)
-              }
+              onClick={() => onClickButton(`/profile/myinfo/edit`)}
             >
               프로필 수정
             </Button>
